requirements-common.txt
# api generation
namex>=0.0.8

# formatting and linting
mypy
ruff

# testing
absl-py
pytest
pytest-cov
numpy

# pip build
build

# required dependencies, should match dependencies in pyproject.toml
# TODO revert. Needed to get keras-nightly until 3.10 is released.
# keras
keras-nightly
ml-dtypes
